Use server components
